+++ main.go
@@ -1,17 +1,43 @@
 package main
 
 import (
+	"flag"
+	"fmt"
+	"os"
+	"strings"
 	"sync"
 
 	"github.com/apiotrowski312/benchHog/results"
 )
 
-const (
-	ratio            = 25
-	numberOfRequests = 250
-)
-
 func main() {
+
+	//
+
+	if len(os.Args) < 2 {
+		fmt.Println("Provide link to site")
+		os.Exit(1)
+	}
+	link := os.Args[1]
+
+	if !strings.HasPrefix(link, "http") {
+		link = "http://" + link
+
+	}
+	// Parse link (add http and e)
+
+	// Flags
+	var numberOfRequests int
+	var ratio int
+
+	description := "Provide number of requests"
+	flag.IntVar(&numberOfRequests, "request", 100, description)
+	flag.IntVar(&numberOfRequests, "r", 100, description+" (shorthand)")
+
+	description = "Provide ratio"
+	flag.IntVar(&ratio, "ratio", 10, description)
+
+	flag.Parse()
 
 	var wg sync.WaitGroup
 
@@ -21,7 +47,7 @@ func main() {
 	wg.Add(numberOfRequests)
 	for i := 0; i < numberOfRequests; i++ {
 		limitRatio <- 1
-		go Get("http://onet.pl", limitRatio, &wg, measurment)
+		go Get(link, limitRatio, &wg, measurment)
 	}
 
 	wg.Wait()