+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 
+## 0.0.2 - under development
+
+- improve release script
+- add flags for ratio and number of requests
+- add position argument to get URL
+
 ## 0.0.1 - 2019-10-04
 ### Added
 - Makefile